Move check into main.rs for logging purposes and rearrange items in main.rs
