+++ testbed/src/helpers.rs
@@ -50,28 +50,6 @@ pub fn subtest(text: &str) {
     *buf = String::from(text);
 }
 
-pub fn check(title: &str, test: TestFn) {
-    info!("Checking {} ...", title);
-    *REASON_TO_FAIL.lock().unwrap() = String::default();
-    *SUBTEST.lock().unwrap() = String::default();
-    TRAPPED_ERROR.store(false, Ordering::SeqCst);
-    SIMULTANEOUS_ERRORS.store(0usize, Ordering::SeqCst);
-    TOTALTESTS.fetch_add(1usize, Ordering::SeqCst);
-    if test().is_ok() && !TRAPPED_ERROR.load(Ordering::SeqCst) {
-        info!("OK");
-    } else {
-        error!("FAILED");
-        let subtest = SUBTEST.lock().unwrap();
-        let reason_to_fail = REASON_TO_FAIL.lock().unwrap();
-        if subtest.len() == 0 {
-            error!("{}: [{}]\n\t{}", title, subtest, reason_to_fail);
-        } else {
-            error!("{}:\n\t{}", title, reason_to_fail);
-        }
-        TOTALFAIL.fetch_add(1, Ordering::SeqCst);
-    }
-}
-
 pub fn clip(v: f64) -> f64 {
     v.clamp(0.0, 1.0)
 }
